PONG_NAME
Pong is a two-player game. players get scores when they hit the ball, as it is a digital game paddles are used to hit the ball instead of tennis bats which are positioned on either side of the screen and players use up and down keys to move the paddles. whoever hits the ball score is incremented.